ADD: Added parse_setup for parsing functions
