+++ include/parser.h
@@ -8,6 +8,9 @@
 # define RGB_FACTOR 0.00392156862745
 
 int		parse_line(char *line, t_scene *scene);
+int		parse_float(char **line, float *value);
+int		parse_vec3(char **line, t_vec3 *v);
+bool	is_valid_color(t_vec3 color);
 
 int		parse_ambient_light(char **line, t_scene *scene);
